// src/app/Models/UserRegister.interface.ts
export interface User {
    User_ID: number;
    User_FirstName: string;
    User_LastName: string;
    User_Departmenr: string;
    User_Designation: string;
    User_DOB: Date;
    User_Gender: string;
    User_EduLevel: string;
    User_PhoneNo: number;
    User_Location: string;
    User_Address: string;
    User_Email: string;
    User_Password: string;
  }
  